Don't fail remove-export in NFS-Ganesha
When dropping an export over DBUS, if there's a problem
with the removal, we can ignore the failure and move on.
If not, the share ends up in an "error_deleting" state.

There's not much that we can do here besides logging
the failure as an exception.

Change-Id: I73c2706c9d889f8d08057a543b2a963740bb3f33
Closes-Bug: #2035572
Signed-off-by: Goutham Pacha Ravi <gouthampravi@gmail.com>
(cherry picked from commit 609b925a51b230c93d8ad408cf8ba351657dd83e)