[Pure Storage] Add support for multiple data VIPs
Minor change to allow multiple data VIP addresses
to be specified in the `flashblade_data_vip` parameter.

Each data VIP will be used to define the file system
export address for shares created. The first address in
the list will be clasified as the Preferred export
address, but this is not enforced.

Change-Id: Ia3bfdd5ce1bfd31d10193d1f222bf27b38c8ef43