sapcc: allow to ensure share instances in 'creating'
The list that is checked has only share instances that already
got a host field set, i.e. the creation stuck at a late stage
(after scheduling). Those should be recoverable here.

In the NetApp driver: always set the status to 'available'.
In case something goes wrong, the share instance is not returned
at all (and therefore not updated) in favor of setting status to
'error', which would be too invasive in certain situation.

Change-Id: I8e25e2bc8880e587457fd4e8cb3fa62868f29f9c