Deprecate Windows SMB share driver
This feature was added and maintained by Cloudbase as part of effort to
support Windows-only OpenStack but Winstackers project was retired and
there seems to be no on requiring this deployment architecture now.

Because the driver hasn't been get any meaningful update for some time
and we no longer have Hyper-V CI active, this proposes deprecating
the driver so that we can retire it in 2024.2 at the earliest.

[1] d9c0bda0b7afa2bf419bd98b819bbb8315e8d530

Change-Id: Icec29f5512d4c4ce8c8d7ccb1f2d3823bffa7ed9