Deprecate the Ceph NFSProtocolHelper
A couple of releases ago, we started supporting clustered NFS
gateways [1], and we mentioned we had clear intentions to make
this the default choice for the deployments.

In this change, we are deprecating the NFSProtocolHelper, which
will be removed in the next skip-level release.

This means that Manila will no longer support exports created over
a NFS Ganesha instance that is not managed by the Ceph Cluster.

[1] https://review.opendev.org/c/openstack/manila/+/848987

Change-Id: I062a8c552d7d94e010cef6694bec6da68d7ea180