[CI] Grenade job changes for 2024.2 release cycle
We're in a non-SLURP release cycle; this means that we'd test upgrading to
the release produced by this cycle is supported only from stable/2024.1 (Caracal).

So set "grenade_from_branch" to stable/2024.1 and run only the "skip-level"
grenade job. When we get into the 2025.1 cycle, we can run the regular
grenade job as well.

Change-Id: If17ddd1bc06810d94db43dec3bb877d5045621ad
Signed-off-by: Goutham Pacha Ravi <gouthampravi@gmail.com>