svm migration across physical networks with multiple port bindings
This patch provides the support for share server migrations between
different physical networks. It is achieved by creating an inactive port
binding on the target host during share server migration, and then cut
it over to target host when migration completes. The implementation is
driver agnostic, so this feature should be not limited to specific
driver, although it is only tested against NetApp driver. The feature is
enabled by a new share service option
'server_migration_extend_neutron_network'. The support for multiple
binding Manila ports is added in Neutron's 2023.1 release.

Closes-Bug: #2002019
Change-Id: I18d0dd1847a09f28989b1a2b62fc6ff8f8155def