[ci] Fix zuul configuration errors
Zuul encountered deprecated syntax while parsing
its configuration for the "manila-grenade-skip-level"
job. The "branch" regex didn't conform to RE2 syntax.

The "branch" annotation isn't necessary because
this repository is branched and zuul is capable of
using the job configuration from a specific branch.
So we can just drop it instead of fixing it.

Change-Id: I12dc07c3a62a7d8ca7b895fd6dc98e165f24e944