Manage/unmanage for CephFS drivers
Implement manage/unmanage for shares and snapshots to the CephFS
drivers.

- Instead of providing the export locations, users will need to
provide the subvolume/subvolume snapshot name in the
export/provider location parameter.

- The size of the snapshots will follow the shares' size, as we
can not currently be sure of the real size of the snapshot. This
must be fixed after the issue is addressed within the ceph
backend [1].

- We will attempt to resize subvolumes whose size are "infinite"
to the size value provided within the driver_options parameter.

- Start storing the provider location of the snapshots after
their creation, so it can be reused later.

[1] https://tracker.ceph.com/issues/67215#change-270669

Partially-Implements: bp manage-unmanage-cephfs-shares-and-snapshots
Change-Id: I5e86d09ea5ba8e61ecf04c52b25c639dfc823737
Signed-off-by: Carlos Eduardo <ces.eduardo98@gmail.com>