Allow clean service shutdown
currently when the parent service that manages workers shuts down,
an ugly trace is produced, as it tries to update its state in DB
to 'stopped' but there is actually no record in DB for this servie
(only worker threads are registered in DB, not the parent thread).

Catch this error and just log a warning the same as already done
in 'kill' method of a service.

Change-Id: Ic4ea3c88c845b46b1d183e71653104bbd08929ff
Related-Bug: #2089534