Add periodic task that detects down services
when a service exits non-gracefully, API will still continue to report
the service state as 'up', as it is only the service itself
(via its manager) who can currently update its state.

This patch adds a periodic task to the manila scheduler, that will
periodically check the list of services, and set service status in DB
as 'down' if the service did not heartbeat within alloted time
and is not in 'stopped' or 'down' state already.

Change-Id: Iefa48ce367a87fd656c91a1ba71bf0e5eaa1e553
Closes-Bug: #2089534