NetApp: delete vlan even if ipspace is reused
Manila keep deleted volumes (and therefore share servers/vservers) for
a while to be able to restore. Those entities live in an ipspace that
is tied to a neutron subnet and fixed vlan ID per NetApp storage
backend as long as a port is bound to that backend. Now during delete
on manila side the port gets released which also cleans the vlan ID
in neutron if it is the last one.

If then a new port in the same subnet is bound to that backend it gets
a new vlan ID, but ends up in the same ipspace. This leads to two vlans
in the same ipspace. When the old volume is ready to be really deleted,
vserver cleanup on manila side tries to delete the ipspace, but won't
because it is still in use and wrongly also did not clean up the vlan.
It assumed wrongly, the vlan would be in use, too. This has been fixed.

The guard, to not attempt to delete the cluster ipspace or ipspaces
that are still in use, has moved to a central place in the cmode client.

Closes-bug: #2097522
Co-authored-by: Maurice Escher <maurice.escher@sap.com>
Change-Id: I84b126eef4c440c0f65bd29419bbdbc1f578e352