Stop allowing generic backups for non-NFS protocols
The share backups feature has a generic approach for share backups
using the manila-data service. The generic approach will mount the
shares and transfer the data over from one share to another, and
in this case, this is done with backups.

However, we currently only coded support for mounting NFS shares,
but we were not checking for the share protocol prior to attempting
the mount, causing the mount operation to fail.

We fix that by checking if hints for third party backup solutions
were added to the backup information. If not, it will use the
generic backup approach and we need to ensure that only NFS shares
are allowed.

Closes-Bug: #2088063
Change-Id: I25cb02fa5e4b87e9adf28f51157522f301b87e32
Signed-off-by: Carlos da Silva <ces.eduardo98@gmail.com>