Make grenade skip level job non-voting temporarily
The manila-grenade-skip-level-always is currently failing while
it attempts to initialize neutron.

This job should focus on testing upgrades from one tick release to
another, so we don't necessarily need to have it as a voting job
on a tock release.

This change temporarily makes the job non-voting for the 2025.2
Flamingo release.

Change-Id: Ibfd20756507b219633a035b58a3ed5082082d467
Signed-off-by: Carlos da Silva <ces.eduardo98@gmail.com>