Support region_name for identity API access
The options in the [barbican] sections are now used to connect to
identity API, to manage application credentials to access encryption
keys. Adds the region_name option to this section so that the code
can look up the appropriate identity endpoint even when service
catalog contains multiple endpoints for multiple regions.

Change-Id: I06894ca9f40a9dc755c7af9dd921dc61735d6b5f
Signed-off-by: Takashi Kajinami <kajinamit@oss.nttdata.com>