Fix shares getting stuck in ensuring status
Shares are getting stuck in ensuring status when the share backend
does not support the bulk ensure shares operation.

At the beginning of the operation we update the shares status but
we never revert the shares' status in case the updated_instances
list is not updated when the backend ensure the shares
indivudually.

This change fixes that behavior by rolling back the status of the
shares individually and also updating the service status when the
operation is done.

Closes-Bug: #2127023
Change-Id: I223b035220788111502f4187dcb75f1d456c72f1
Signed-off-by: Carlos da Silva <ces.eduardo98@gmail.com>
(cherry picked from commit 795ff69c55ea686d2918c58f9e99dff5114c2e88)