Remove the confusing 'deferred deletion' wording from the user guide
The 'Resource locks' section of the 'Create and manage shares' user guide refers to "deferred deletion" when describing the recycle bin (soft deletion) feature.
This is misleading, since 'deferred deletion' is a different configuration feature controlled by the 'is_deferred_deletion_enabled' option.

This patch updates the documentation to remove the reference to "deferred deletion" to prevent user confusion.

Closes-Bug: #2110220: [DOC] distinguish 'recycle bin' and 'deferred deletion'.
Signed-off-by: Rose Kimondo <rosewairimu1@gmail.com>
Change-Id: Ied675b1bb4f84d6ef768addb1738a1e3a5c577d7