Fix rally jobs
Previously, we forced rally jobs to do not use the global venv when
it was introduced a couple of releases ago, as it caused a breakage.

Later on, we started seeing a failure in the rally jobs, where
os-testr installation/upgrade failed, and the reason was that we
were attempting to upgrade a package that was externally-managed,
indicating that this installation/upgrade should be within the
venv.

This change switches the rally jobs to using the global venv,
as the issues we saw back then were fixed.

Change-Id: I33275c6eb5b137d78b10a9942ac4f765d011eac1
Signed-off-by: Carlos da Silva <ces.eduardo98@gmail.com>